初步实现tokenwise和layerwise的GPU冷启动demo

+++ benchmarks/benchmark_utils.py
@@ -39,7 +39,7 @@ def _warmup_inference():
     print("Warming up inference")
     model_name = "facebook/opt-6.7b"
     model = AutoModelForCausalLM.from_pretrained(
-        model_name, torch_dtype=torch.float16, device_map="auto"
+        model_name, torch_dtype=torch.bfloat16, device_map="auto"
     )
     prompts = [
         "The quick brown fox jumps over the lazy dog.",
@@ -95,7 +95,7 @@ def measure(
                 f"{model_name}_{model_idx}",
                 storage_path=model_dir,
                 device_map="auto",
-                torch_dtype=torch.float16,
+                torch_dtype=torch.bfloat16,
             )
             end_time = time.time()
         elif model_format == "safetensors":
@@ -105,7 +105,7 @@ def measure(
             start_time = time.time()
             model = AutoModelForCausalLM.from_pretrained(
                 model_path,
-                torch_dtype=torch.float16,
+                torch_dtype=torch.bfloat16,
                 device_map="auto",
             )
             end_time = time.time()
