+++ tests/new_cli_test/test_deploy.py
@@ -32,7 +32,7 @@ class TestDeployCommand(unittest.TestCase):
             "backend_config": {
                 "pretrained_model_name_or_path": "",
                 "device_map": "auto",
-                "torch_dtype": "float16",
+                "torch_dtype": "bfloat16",
             },
         }
         user_config = {
