+++ tests/replay_test/deploy_dummy_models.py
@@ -36,7 +36,7 @@ config_template = {
     "backend_config": {
         "pretrained_model_name_or_path": "",
         "device_map": "auto",
-        "torch_dtype": "float16",
+        "torch_dtype": "bfloat16",
     },
 }
